+++ dhis-2/dhis-services/dhis-service-reporting/src/test/java/org/hisp/dhis/validation/EventValidationServiceTest.java
@@ -33,6 +33,7 @@ import org.hisp.dhis.DhisTest;
 import org.hisp.dhis.IntegrationTest;
 import org.hisp.dhis.analytics.AggregationType;
 import org.hisp.dhis.analytics.AnalyticsService;
+import org.hisp.dhis.category.CategoryManager;
 import org.hisp.dhis.common.DimensionalObject;
 import org.hisp.dhis.common.Grid;
 import org.hisp.dhis.common.GridHeader;
@@ -129,6 +130,9 @@ public class EventValidationServiceTest
     private ValidationRuleService validationRuleService;
 
     private DataElementCategoryOptionCombo defaultCombo;
+
+    @Autowired
+    private CategoryManager categoryManager;
 
     private OrganisationUnit orgUnitA;
 
@@ -233,7 +237,7 @@ public class EventValidationServiceTest
         programStageInstanceService.addProgramStageInstance( stageInstanceA );
         programStageInstanceService.addProgramStageInstance( stageInstanceB );
 
-        categoryService.addAndPruneAllOptionCombos();
+        categoryManager.addAndPruneAllOptionCombos();
 
         Expression expressionA = new Expression( EXPRESSION_A, "ProgramTrackedEntityAttribute" );
         Expression expressionD = new Expression( EXPRESSION_D, "ProgramDataElement" );
